requirements.txt
flask==3.0.0
werkzeug==3.0.1
psycopg2-binary==2.9.9
reportlab==4.1.0
pytest==8.1.1
pytest-html==4.1.1
pytest-cov==5.0.0
flake8==7.0.0
